vmm: Wait for the completion of PCI device removal
Whenever a PCI device was unplugged from the VM, the HTTP thread was
returning right after the internal logic was updated and the ACPI
notification was triggered. This means at this point, the OS didn't
fully removed the device, which also means the function eject_device()
in the DeviceManager wasn't invoked yet. Since this function performs
all the necessary cleaning required to really consider the device as
removed, it's important to wait for this codepath to be executed before
to let the HTTP request return.

By relying on a mechanism of barriers, the vmm thread (from the HTTP
request) and the vCPU thread (from the OS triggering the eject of the
device) have a way to wait for each other. Especially, this allows the
vmm thread to know when the removal is complete, which avoid potential
race conditions if someone wants to add a new device right after this.

Fixes #4038

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>