vmm: Use the identifier generator for single devices
There are multiple device types for which a single instance will be
created. The list contains iommu, ioapic, gpio, serial, console, rng,
balloon and watchdog devices. So far, the identifier associated with
each device was hardcoded, but this prevents the VMM from ensuring the
uniqueness of each identifier.

This patch relies on the identifier generator from DeviceManager to
create unique identifiers for these devices, relying on the list of
existing devices to prevent from conflicts.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>