device: add fw_cfg device implementation
Firmware Configuration device derives from Qemu used for pass some info
from VMM to firmware. To let firmware boot kernel specified by "-kernel"
we must let firmware be aware of it that we choose fw_cfg to do this. Also
we can reuse the device driver in edk2.

As only part of the function of fw_cfg device will be used, we only
implement part of this device to minimize the size of it.
More info see [1].

[1] https://github.com/qemu/qemu/blob/master/docs/specs/fw_cfg.rst

Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>