Initial PCIE capability filtering
This PR encodes an unsupported blocklist of PCIE capabilities for
VFIO. This is needed because some PCIE devices pass through SRIOV
capabilities, as well as other capabilities, which are not
supported by the VFIO API or cloud-hypervisor. The blocklist
will simply block the capabilities from being exposed in the guest
during PCIE configuration.

Signed-off-by: Steven Dake <sdake@lambdal.com>