option_parser: For tuple list type strip brackets only once
The new option parsing for "[..,]" syntax is much stricter and requires
that everything is balanced in terms of brackets. Change the stripping
of the brackets to only strip the first and last one rather than
multiple:

e.g. "[[a,b], [c,d]]" becomes "[a,b],[c,d]" rather than "a,b], [c"

Signed-off-by: Rob Bradford <robert.bradford@intel.com>